Remove: api/history.ts(불필요한 모듈화 삭제)
